requirement.txt
absl-py==1.4.0
array-record==0.5.0
asttokens==2.4.1
astunparse==1.6.3
bleach==6.1.0
blinker==1.7.0
cachetools==5.3.3
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
contourpy==1.2.0
cycler==0.12.1
Cython==3.0.9
dataclasses==0.6
decorator==5.1.1
dm-tree==0.1.8
esdk-obs-python==3.23.12
etils==1.7.0
exceptiongroup==1.2.0
executing==2.0.1
filelock==3.13.1
Flask==3.0.2
flatbuffers==23.5.26
fonttools==4.49.0
fsspec==2024.2.0
gast==0.4.0
gin-config==0.5.0
google-api-core==2.17.1
google-api-python-client==2.121.0
google-auth==2.28.1
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.0.0
google-cloud-bigquery==3.18.0
google-cloud-core==2.4.1
google-crc32c==1.5.0
google-pasta==0.2.0
google-resumable-media==2.7.0
googleapis-common-protos==1.62.0
grpcio==1.62.0
h5py==3.10.0
httplib2==0.22.0
huggingface-hub==0.21.4
idna==3.6
immutabledict==4.1.0
importlib_resources==6.1.2
install==1.3.5
ipython==8.22.2
itsdangerous==2.1.2
jedi==0.19.1
Jinja2==3.1.3
joblib==1.3.2
kaggle==1.6.6
keras==2.13.1
Keras-Preprocessing==1.1.2
kiwisolver==1.4.5
libclang==16.0.6
lxml==5.1.0
Markdown==3.5.2
MarkupSafe==2.1.5
matplotlib==3.8.3
matplotlib-inline==0.1.6
ml-dtypes==0.2.0
numpy==1.24.3
nvidia-cublas-cu12==12.4.2.65
nvidia-cuda-nvrtc-cu12==12.4.99
nvidia-cuda-runtime-cu12==12.4.99
nvidia-cudnn-cu12==8.9.7.29
oauth2client==4.1.3
oauthlib==3.2.2
opencv-python-headless==4.9.0.80
opt-einsum==3.3.0
packaging==23.2
pandas==2.2.1
parso==0.8.3
pexpect==4.9.0
pillow==10.2.0
portalocker==2.8.2
promise==2.3
prompt-toolkit==3.0.43
protobuf==3.20.3
psutil==5.9.8
ptyprocess==0.7.0
pure-eval==0.2.2
py-cpuinfo==9.0.0
pyasn1==0.5.1
pyasn1-modules==0.3.0
pycryptodome==3.20.0
Pygments==2.17.2
pyparsing==3.1.1
python-dateutil==2.9.0.post0
python-slugify==8.0.4
pytz==2024.1
PyYAML==6.0.1
regex==2023.12.25
requests==2.31.0
requests-oauthlib==1.3.1
rsa==4.9
sacrebleu==2.4.0
safetensors==0.4.2
scikit-learn==1.4.1.post1
scipy==1.12.0
sentencepiece==0.2.0
seqeval==1.2.2
six==1.16.0
stack-data==0.6.3
tabulate==0.9.0
tensorboard==2.13.0
tensorboard-data-server==0.7.2
tensorboard-plugin-wit==1.8.1
tensorflow==2.13.1
tensorflow-addons==0.23.0
tensorflow-datasets==4.9.4
tensorflow-estimator==2.13.0
tensorflow-hub==0.16.1
tensorflow-io-gcs-filesystem==0.36.0
tensorflow-metadata==1.14.0
tensorflow-model-optimization==0.8.0
tensorflow-text==2.13.0
tensorrt==8.6.1.post1
tensorrt-bindings==8.6.1
termcolor==2.4.0
text-unidecode==1.3
tf-keras==2.15.0
tf-models-official==2.13.2
tf-slim==1.1.0
threadpoolctl==3.3.0
tokenizers==0.15.2
toml==0.10.2
tqdm==4.66.2
traitlets==5.14.1
transformers==4.38.2
typeguard==2.13.3
typing_extensions==4.5.0
tzdata==2024.1
uritemplate==4.1.1
urllib3==2.2.1
wcwidth==0.2.13
webencodings==0.5.1
Werkzeug==3.0.1
wrapt==1.14.1
zipp==3.17.0
